build(deps): restrict b13/container version to < 3.1.2
* Update composer.json

Container got updatet in 3.1.2 they redo the ContainerProcessor and it breaks the headless-container Processor. 
To prevent breaking in production. I add a strict Version constrain to the composer.json.

after that we should update. we should go to v3 with headless-container and depend on container > 3.1.2 to make clear that here is a breaking change.

* Update composer.json

Co-authored-by: Ramón <schlosser@itplusx.de>

---------

Co-authored-by: Ramón <schlosser@itplusx.de>